[SPARK-25534][SQL] Make `SupportWithSQLConf` trait
